move import sanitization utils to `@/lib/autogpt-server-api/utils`
